Moves default value into __init__ default values.
Don't know why I didn't do this in the first place, had some silly idea that it
looked like it was a long string to have in a function header/people might want
to see a value like this listed prominently in the code layout.

Thought: change api_key to _api_key? Really shouldn't get used outside this
object.